Refacto: change Encoder and Decoder modules into builder functions Moving final_layer builder into decoder.py for consistency Create: build_full_decoder function to build the full stack Decoder part.
